recommendation from dataset

+++ MovieRS.API/Models/DetailGenre.cs
@@ -7,5 +7,5 @@ public partial class DetailGenre
 {
     public int IdMovie { get; set; }
 
-    public int IdGenre { get; set; }
+    public byte IdGenre { get; set; }
 }
